backend/pyproject.toml
[tool.poetry]
name = "core-backend"
version = "0.1.0"
description = ""
authors = ["uuip <uuip@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
requests = "^2.28.1"
web3 = "^5.30.0"
djangorestframework-simplejwt = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
